Move compiler dialog functions from edit_project module.
